python/requirements.txt
pandas
requests
sqlalchemy
psycopg2-binary
odfpy
requests
